Refactor machines and roles
Previously `machines` were tightly-coupled to the computer's hostname
and software rather than the hardware. Now each machine entry describes
(mostly) the physical machine (or, e.g., the fact that it needs to dual
boot another OS), and `roles` couple a hostname to a collection of
software and settings to be used on that hardware.

In theory this means that if I want to have work apps temporarily
installed on my home machine, I can just switch its role for a bit.