The great profiles refactor, part one
I added some new packages during a game jam and realised that my
`profiles` directory is sort-of misnamed, and `profiles` isn't really
a name I love for this concept. This commit marks the beginning of a
refactor to fix that.

Here I just create explicit "gamedev" and "social" config definitions.